Introduce a new config to manage rc type change functionality:
Enabling this function leads some STBs to bootloop although they have the /proc/stb/ir/rc/type path so that check isn't enough and also why block this feature on some hardwares when we can disable the function by default and let users manage it.

So now we check the requirement which is /proc/stb/ir/rc/type and let users to enable it if they have the requirement.

Also disable HDMI record for SH4 as there's no HDMI-In