Add new information system
* [Information.py] Add new information system

The Information.py code replaces the About.py code.  This is a more optimised and configurable information display system.  This version is more comprehensive.  The UI supports information updates without needing to exit and reload the screen.  The code is designed to promptly provide the user interface and update information that takes some time to access or generate.

* [menu.xml] Update menus for new Information screens