[Directories.py] Add fileWriteLine() and fileWriteLines()
The two new methods create a central library for Enigma2 code to write data to disk files.

There is a lot of code that that does a open(filename, "r").write("text"). This code generally offers no protection against errors and crashes. These new methods perform a similar function but errors are reported if the file operations fail.

If debug mode is enabled the execution stack is used to nominate the calling code module and the line in that module that called these methods. This should allow debugging when required. Debugging should not be used by default as this will add significant time to the code execution (while the program stack is processed).