[ScrollLabel.py] Propose stop modifying text
This change is a proposal to stop ScrollLabel from modifying text before display.  At the moment the left column is right stripped and the right column is left stripped.  This can cause issues if the text to be displayed is intended to be indented with spaces.

This proposal removes the strip() method calls with the aim of seeing if any usage of ScrollLabel is supplying white space that is undesirable.  My suggestions is that if code is supplying incorrect text then the source be fixed rather than forcing all text to be stripped.