[Pixmap.py] Add callback option for movement
This change adds an optional callback method to be run when the moving animation completes.  The "callback" parameter to the startMoving() method is completely optional but if specified must be a callable method.  If it is not callable it will be silently ignored.

This new functionality will initially be used by the new remote control help system.