[Directories.py] Improve isPluginInstalled method
This change combines the four separate and related plugin location methods into a single method.  The new code can also check nonstandard plugin directories, if they are used.

The previous ability to check if a plugin was in Extensions or in SystemPlugins is preserved with the optional pluginType argument.  If pluginType is not specified then all plugin directories are checked.