Refactor remote control and help system
* Refactor remote control and help system

* [HelpMenu.py] Remove some test code

* InputDevice.py] Imnprove getOpenWebifHTML() method

If a valid OpenWebif HTML template can not be created from the data in the remote control XML files then return None so that OpenWebif can use the old method of obtaining the remote control HTML mapping.

* [InputDevice.py] Set default label to name

This fix creates a button label using the name attribute if the label attribute is not defined in the remote control XML file.

* [Makefile.am] Demove defunct Keyboard.py

* [Makefile.am] Remove defunct keymapparser.py

* [PluginComponent.py] Remove defunct keymapparser import

Replace the import of the defunct keymapparser.py import with its new replacement of loadKeymap() in Components.ActionMap.py.

* [SystemInfo.py] Transfer more SystemInfo items to BoxInfo

* [HelpMenu.py] Add correction to support Wizards

This is a temporary change to support existing wizards that use the old button names.  The fix is to convert the old names into the new keyId numbers.

* [InputDevice.py] Correct some BoxInfo variable names