SoftcamSetup.py: use with user control method
NoSave method in this setup with green button action, it is not the best option.

With this change the configuration control is complete, the green button will save the selection.

by default cardserver and softcam is None.
Restart options defaults to None.

Any change that is made to the config if we save with the green button will keep it, in this way we can preserve the user preferences.

Example:
If a user has a non-default configuration saved in his copy, it will be kept in the restoration, without having to reconfigure this setup to start his softcam