Update all remote control XML definitions
This change removes the "name=" attribute and data to be replaced with the "id=" attribute.  This attribute creates a direct match between the remote control definition file and the key code emitted by the driver.  The attribute name "id=" and data was chosen as it directly matches the "id=" attribute in the keymap.xml file.  Both files use the "KEY_xyz" key names as defined in keyids.py.

This change also sees the removal of the KeyBindings.py file as it is no longer required.